Replace DAI with WBTC
Public interfaces are not renamed to avoid unnecessary breaking changes.
Token contract addresses are changed to reflect WBTC.

Nectar's math treats WBTC as if it had 18 decimals although in reality
it only has 8. This is required so that we can represents decimal rates.

As a nice side-effect, we can remove most of the float-math stuff in
nectar.